- Review 좋아요/싫어요 기능 추가  - Review 평가시 작성자의 점수 변경  - Review 중복 평가 방지용 like_history 엔티티 추가 -> 해당 테이블에 기록이 없을시에만 평가 가능  - Menu에 url 필드 추가
TODO:
 - 수정/삭제에 대한 이미지 처리
 - Review 반환 시 이미지 동반 반환

+++ LetsEatAll/src/main/java/com/letseatall/letseatall/data/Entity/Restaurant.java
@@ -20,7 +20,7 @@ public class Restaurant {
     @Column(nullable = false)
     private String name;
     @Column(nullable = false)
-    private int score;
+    private double score;
     @Column
     private String addr;
     @ManyToOne
